+++ src/array/index.ts
@@ -1,2 +1,3 @@
 export * from './array.js'
 export * from './items.js'
+export * from './maxItems.js'