+++ src/array/maxItems.ts
@@ -5,13 +5,14 @@ import { ruleYobta } from '../ruleYobta/index.js'
 export const maxItemsMessage = (limit: number): string =>
   `It should be within ${pluralizeEn(limit, 'item')}`
 
-export const maxItemsYobta = <I extends unknown[]>(
-  limit: number,
+export const maxItems = <I extends unknown[]>(
+  limit: () => number,
   message = maxItemsMessage,
 ): YobtaSyncRule<I, I> =>
   ruleYobta<I, I>(input => {
-    if (input.length > limit) {
-      throw new Error(message(limit))
+    const l = limit()
+    if (input.length > l) {
+      throw new Error(message(l))
     }
 
     return input